requirements.txt
beautifulsoup4==4.5.1
cffi==1.7.0
dnspython==1.15.0
docutils==0.12
Flask==0.11.1
Flask-Cors==2.1.2
gevent==1.1.2
googlefinance==0.7
googlemaps==2.4.5
Jinja2==2.8
jsonschema==2.5.1
oauth2==1.9.0.post1
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.3.0
Pygments==2.1.3
pylint==1.5.4
pyOpenSSL==16.0.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.5.3
python-dotenv==0.6.2
requests==2.11.1
responses==0.5.1
Sphinx==1.4.6
yelp==1.0.2